--- packages/earth-map/src/components/index-content/styles.scss
@@ -1,159 +1,0 @@
-/*
-  Copyright 2018-2020 National Geographic Society
-
-  Use of this software does not constitute endorsement by National Geographic
-  Society (NGS). The NGS name and NGS logo may not be used for any purpose without
-  written permission from NGS.
-
-  Licensed under the Apache License, Version 2.0 (the "License"); you may not use
-  this file except in compliance with the License. You may obtain a copy of the
-  License at
-
-      https://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software distributed
-  under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-  CONDITIONS OF ANY KIND, either express or implied. See the License for the
-  specific language governing permissions and limitations under the License.
-*/
-
-@import '~styles/config';
-
-.c-index-content {
-  overflow: auto;
-  -webkit-overflow-scrolling: touch;
-  color: $marapp-gray-1;
-
-  &.-hidden {
-    visibility: hidden;
-  }
-
-  &.-active {
-    z-index: $z-index-2;
-  }
-
-  .index-content--item {
-    cursor: pointer;
-    position: relative;
-    outline: none;
-
-    > * {
-      margin: $space-1/2 0 0;
-    }
-
-
-    &.-horizontal {
-      display: flex;
-      justify-content: flex-start;
-      align-items: center;
-      padding-right: $space-1 * 3;
-      margin: $space-1 * 2 0 0;
-
-      &:first-child {
-        margin: 0;
-      }
-
-      > * {
-        margin: 0 0 0 $space-1;
-
-        &:first-child {
-          margin: 0;
-        }
-      }
-    }
-
-    .index-content--item-title {
-      font-size: 9px;
-      font-weight: $font-weight-medium;
-      letter-spacing: 1.5px;
-      text-transform: uppercase;
-    }
-
-    .index-content--item-bg {
-      position: relative;
-      width: 100%;
-      height: 40px;
-      background: $marapp-gray-100;
-      background-size: cover;
-      background-position: center;
-      border: 1px solid $marapp-gray-2;
-
-      &:after {
-        content: "";
-        position: absolute;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-        border: 2px solid transparent;
-      }
-    }
-
-    .index-content--item-switch {
-      position: relative;
-      width: 20px;
-      height: 7px;
-      background: $marapp-gray-2;
-      border-radius: 7px;
-      transition: background $animation-time linear;
-      flex-shrink: 0;
-
-      span {
-        display: block;
-        position: absolute;
-        top: 50%;
-        left: 0;
-        transform: translate(0%,-50%);
-        width: 9px;
-        height: 9px;
-        border-radius: 50%;
-        background: $marapp-gray-9;
-        box-shadow: 0 0 0 2px $marapp-gray-0;
-        transition: all $animation-time $ease-in-out-sine;
-      }
-    }
-
-    .index-content--item-info {
-      position: absolute;
-      right: 0;
-      top: 0;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      width: 14px;
-      height: 14px;
-      background: $marapp-gray-0;
-      fill: $marapp-gray-100;
-      border-radius: 50%;
-
-      svg {
-        display: block;
-      }
-    }
-
-    &:hover {
-      .index-content--item-bg {
-        &:after {
-          border-color: rgba($marapp-primary-color, 0.5);
-        }
-      }
-    }
-
-    &.-active {
-      .index-content--item-bg {
-        &:after {
-          border-color: $marapp-primary-color;
-        }
-      }
-
-      .index-content--item-switch {
-        background: #006EFF;
-
-        span {
-          left: 100%;
-          transform: translate(-100%, -50%);
-        }
-      }
-    }
-  }
-}
