+++ packages/earth-map/src/pages-client/earth/collection-details/styles.scss
